@@ -12,42 +12,51 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.1",
-    "@angular/animations": "^4.3.6",
-    "@angular/cdk": "^2.0.0-beta.10",
-    "@angular/common": "^4.2.4",
-    "@angular/compiler": "^4.2.4",
-    "@angular/core": "^4.2.4",
-    "@angular/forms": "^4.2.4",
-    "@angular/http": "^4.2.4",
-    "@angular/material": "^2.0.0-beta.10",
-    "@angular/platform-browser": "^4.2.4",
-    "@angular/platform-browser-dynamic": "^4.2.4",
-    "@angular/router": "^4.2.4",
-    "core-js": "^2.4.1",
-    "ngx-bootstrap": "^1.9.1",
-    "rxjs": "^5.4.2",
-    "zone.js": "^0.8.14"
+    "@agm/core": "^3.0.0-beta.0",
+    "@angular-devkit/core": "^10.1.0",
+    "@angular/animations": "^10.1.0",
+    "@angular/cdk": "^10.2.0",
+    "@angular/common": "^10.1.0",
+    "@angular/compiler": "^10.1.0",
+    "@angular/core": "^10.1.0",
+    "@angular/forms": "^10.1.0",
+    "@angular/http": "^7.2.16",
+    "@angular/material": "^10.2.0",
+    "@angular/platform-browser": "^10.1.0",
+    "@angular/platform-browser-dynamic": "^10.1.0",
+    "@angular/router": "^10.1.0",
+    "bootstrap": "^4.5.2",
+    "core-js": "^3.6.5",
+    "font-awesome": "^4.7.0",
+    "jquery": "^3.5.1",
+    "ngx-bootstrap": "^6.1.0",
+    "popper.js": "^1.16.1",
+    "rxjs": "^6.6.3",
+    "rxjs-compat": "^6.6.3",
+    "tslib": "^2.0.0",
+    "zone.js": "^0.11.1"
   },
   "devDependencies": {
-    "@angular/cli": "1.3.2",
-    "@angular/compiler-cli": "^4.2.4",
-    "@angular/language-service": "^4.2.4",
-    "@types/jasmine": "~2.5.53",
+    "@angular-devkit/build-angular": "~0.1001.0",
+    "@angular/cli": "^10.1.0",
+    "@angular/compiler-cli": "^10.1.0",
+    "@angular/language-service": "^10.1.0",
+    "@types/jasmine": "^3.5.14",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "@types/node": "^14.6.4",
+    "@types/googlemaps": "^3.39.13",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "^3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "^5.2.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-cli": "^2.0.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "^9.0.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.2"
   }
 }
