@@ -3,10 +3,10 @@ import { Injectable } from '@angular/core';
 
 import { Http, Response } from '@angular/http';
 
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
+import { Observable } from 'rxjs';
 import 'rxjs/add/operator/toPromise';
+import { map } from "rxjs/operators"; 
+import { catchError } from "rxjs/operators"; 
 
 @Injectable()
 export class MiHttpService {
@@ -34,9 +34,7 @@ export class MiHttpService {
 
   public httpGetO ( url: string): Observable<Response>
   {
-    return this.http.get( url )
-      .map( ( res: Response ) => res.json())
-      .catch( ( err: any ) => Observable.throw(err.json().error || 'Server error'));
+    return this.http.get( url ).pipe(map( ( res: Response ) => res.json())).pipe(catchError( ( err: any ) => Observable.throw(err.json().error || 'Server error')));
   }
 
 
