@@ -1,11 +1,13 @@
 import { Injectable } from '@angular/core';
 
 import {Http ,Response} from '@angular/http';
+
+import {Observable} from 'rxjs';
+
 import 'rxjs/add/operator/toPromise';
 
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+/*import 'rxjs/add/operator/map';
+import 'rxjs/add/operator/catch';*/
 
 @Injectable()
 export class MiHttpService {
